+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Search dashboard cards by their URLs.
+
 ## [0.1.11] - 2026-08-23
 
 ### Changed